+++ renderer/framebuffer.cc
@@ -95,8 +95,8 @@ Texture Framebuffer::CreateAttachmentTexture(const FramebufferAttachment& attach
   } else if (attachment.texture_meta.gl_internal_format == GL_RG32F) {
     std::vector<glm::vec2> data(texture_meta.data_size_in_byte() / sizeof(glm::vec2), glm::vec2(0, 0));
     return CreateTexture2D(texture_meta, {(data.data())});
-  } else if (attachment.texture_meta.gl_internal_format == GL_R32F) {
-    std::vector<float> data(texture_meta.data_size_in_byte() / sizeof(float), 0);
+  } else if (attachment.texture_meta.gl_internal_format == GL_R32UI) {
+    std::vector<unsigned int> data(texture_meta.data_size_in_byte() / sizeof(unsigned int), 0);
     return CreateTexture2D(texture_meta, {(data.data())});
   } else {
     CGCHECK(false) << " Unsupported Internal format"
